Add guardian-dev.nix
This file provides a wrapper around mkShellNoCC that can be used in a
shell.nix and (in return for a definition of the commands in a project
and the project’s name) gives the user a command for running everything
necessary for the project in tmux.

It’s a draft of a development environment we could reuse across multiple
projects, and I’m adding it here to test fetching it from github.