Add run-all-background command
This commit adds a new command run-all-background to the dev
environment, which runs each of the given commands in turn, sending all
but the last to the background with “&”. This matches a pattern I’ve
seen in start scripts across our repositories.

Mostly, I’m adding this to test depending on a tag: I’ll update the v1
tag to point to this commit and see how depending on the tag in
https://github.com/guardian/editorial-tools-production-monitoring/commit/33a4ff64ddf65408a86e8cb6e4e90969993e4357
handles the update.