Add extraInputs attribute to devEnv
It’s useful for the nix shell to provide other dependencies (e.g. metals
for scala editor integrations) that are not run as part of starting the
app. This commit adds an `extraInputs` attribute which enables this.

Because the attribute is optional, this should maintain backwards
compatibility with v1.0 and v1.1.